bet_writer_consumer/requirements.txt
kafka_python==2.0.2
uvicorn==0.20.0
psycopg==3.1.4


